Fix broken links with Jekyll Relative Links
Multiple links in the guides had incorrect relative URLs when rendered
with GitHub Pages. At the same time, fixing the links for GitHub Pages
would result in incorrect URLs when viewing the guides as Markdown files
on github.com.

Both problems are solved by using the Jekyll Relative Links plug-in [1]
and directly linking to the Markdown files with their .md extension.
With this plug-in, Jekyll automatically makes links to the Markdown
files point to the correct HTML-rendered locations. At the same time,
the links on github.com are still working, because they can now point to
the actual file locations.

Note that the Jekyll Relative Links plug-in is enabled by default on all
GitHub Pages builds. The idea stems from Ben Balter at GitHub, who also
wrote a blog post on that topic [2].

This commit adds the necessary piece of configuration to make use of the
Jekyll Relative Links plug-in. Note that it explicitly needs to be
turned on for collections—otherwise, it wouldn’t work within the guides.
Additionally, this commit makes all guide links point to the respective
Markdown files.

[1] https://github.com/benbalter/jekyll-relative-links
[2] https://github.com/blog/2290-relative-links-for-github-pages